dev-lv2

+++ src/main/java/org/example/scheduleadvanced/service/ScheduleService.java
@@ -5,6 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.example.scheduleadvanced.dto.ScheduleResponseDto;
 import org.example.scheduleadvanced.entity.Schedule;
 import org.example.scheduleadvanced.entity.Member;
+import org.example.scheduleadvanced.exception.UnauthorizedException;
 import org.example.scheduleadvanced.repository.ScheduleRepository;
 import org.example.scheduleadvanced.repository.MemberRepository;
 import org.springframework.http.HttpStatus;
@@ -20,9 +21,9 @@ public class ScheduleService {
     private final ScheduleRepository scheduleRepository;
     private final MemberRepository memberRepository;
 
-    public ScheduleResponseDto createSchedule(String title, String content, long userId){
+    public ScheduleResponseDto createSchedule(String title, String content, long memberId){
         Schedule schedule = new Schedule(title, content);
-        Optional<Member> optionalUser = memberRepository.findById(userId);
+        Optional<Member> optionalUser = memberRepository.findById(memberId);
         Member member = optionalUser.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "사용자를 찾을 수 없습니다."));
         schedule.setMember(member);
         scheduleRepository.save(schedule);
@@ -51,13 +52,22 @@ public class ScheduleService {
     }
 
     @Transactional
-    public ScheduleResponseDto modifySchedule(Long id, String title, String content){
+    public ScheduleResponseDto modifySchedule(Long id, String title, String content, Long memberId){
+        // 자신의 댓글인지 확인
         Schedule schedule = scheduleRepository.findScheduleById(id);
+        if(!schedule.getMember().getId().equals(memberId))
+            throw new UnauthorizedException("자신의 일정만 수정할 수 있습니다.");
+
         schedule.updateSchedule(title, content);
-        return getScheduleById(id);
+        return ScheduleResponseDto.toDto(schedule);
     }
 
-    public void deleteSchedule(Long id){
+    public void deleteSchedule(Long id, Long memberId){
+        // 자신의 댓글인지 확인
+        Schedule schedule = scheduleRepository.findScheduleById(id);
+        if(!schedule.getMember().getId().equals(memberId))
+            throw new UnauthorizedException("자신의 일정만 삭제할 수 있습니다.");
+
         scheduleRepository.delete(scheduleRepository.findScheduleById(id));
     }
 }
